Omorganisering
Flytter kode som hører til en enkelt side inn i app-foldern
