+++ src/api/fetch/saksdetaljer/fetchSoknadsdetaljer.ts
@@ -1,8 +1,8 @@
 import * as R from "remeda";
 
-import { SaksListeResponse } from "../../../generated/model";
-import { getSaksDetaljer } from "../../../generated/ssr/saks-oversikt-controller/saks-oversikt-controller";
-import { SaksDetaljerResponse } from "../../../generated/ssr/model";
+import { SaksListeResponse } from "@generated/model";
+import { getSaksDetaljer } from "@generated/ssr/saks-oversikt-controller/saks-oversikt-controller";
+import { SaksDetaljerResponse } from "@generated/ssr/model";
 
 const fetchSoknadsdetaljer = (saker: SaksListeResponse[]): Promise<SaksDetaljerResponse>[] =>
     // TODO: Filteret her tror jeg ikke trengs, da fiksDigisosId alltid skal være satt. Se TODO i innsyn-api.
