+++ src/api/ssr/authenticatedFetch.ts
@@ -1,6 +1,6 @@
 import { cookies, headers } from "next/headers";
 
-import { getServerEnv } from "../../config/env";
+import { getServerEnv } from "@config/env";
 
 const getAuthorizationHeader = async (): Promise<string | null> => (await headers()).get("authorization");
