+++ src/app/[locale]/landingsside/_components/snarveier/KlagerSnarvei.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { getTranslations } from "next-intl/server";
 import { GavelIcon } from "@navikt/aksel-icons";
 
-import StatusCard from "../soknaderList/statusCard/StatusCard";
+import StatusCard from "@components/statusCard/StatusCard";
 
 const KlagerSnarvei = async () => {
     const t = await getTranslations("Snarveier");
