+++ src/app/[locale]/landingsside/_components/snarveier/UtbetalingerSnarvei.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { BankNoteIcon } from "@navikt/aksel-icons";
 import { getTranslations } from "next-intl/server";
 
-import StatusCard from "../soknaderList/statusCard/StatusCard";
+import StatusCard from "@components/statusCard/StatusCard";
 
 const UtbetalingerSnarvei = async () => {
     const t = await getTranslations("Snarveier");
