+++ src/app/[locale]/landingsside/page.tsx
@@ -4,11 +4,12 @@ import { getTranslations } from "next-intl/server";
 
 import OkonomiskSosialhjelpIcon from "@components/ikoner/OkonomiskSosialhjelp";
 import { getFlag, getToggles } from "@featuretoggles/unleash";
-import Snarveier from "@components/snarveier/Snarveier";
 import NyttigInformasjon from "@components/nyttigInformasjon/NyttigInformasjon";
-import AktiveSoknader from "@components/aktiveSoknader/AktiveSoknader";
 import SokButton from "@components/sokButton/SokButton";
 import ClientBreadcrumbs from "@components/breadcrumbs/ClientBreadcrumbs";
+import AktiveSoknader from "@components/aktiveSoknader/AktiveSoknader";
+
+import Snarveier from "./_components/snarveier/Snarveier";
 
 const Page = async () => {
     const toggle = getFlag("sosialhjelp.innsyn.ny_landingsside", await getToggles());
