+++ src/app/[locale]/soknad/[id]/_components/oversikt/useSteps.tsx
@@ -1,7 +1,7 @@
 import { useParams } from "next/navigation";
 import * as R from "remeda";
 
-import { useHentHendelserBetaSuspense } from "../../../generated/hendelse-controller/hendelse-controller";
+import { useHentHendelserBetaSuspense } from "@generated/hendelse-controller/hendelse-controller";
 
 import SendtStep from "./steps/SendtStep";
 import MottattStep from "./steps/MottattStep";
