+++ src/app/[locale]/soknad/[id]/page.tsx
@@ -1,11 +1,12 @@
 import { notFound } from "next/navigation";
 
 import { getFlag, getToggles } from "@featuretoggles/unleash";
-import { StatusSendtPage } from "@components/statusPage/StatusSendtPage";
 import { hentSoknadsStatus } from "@generated/ssr/soknads-status-controller/soknads-status-controller";
-import { StatusMottattPage } from "@components/statusPage/StatusMottattPage";
-import { StatusUnderBehandlingPage } from "@components/statusPage/StatusUnderBehandlingPage";
-import { StatusFerdigbehandletPage } from "@components/statusPage/StatusFerdigbehandletPage";
+
+import { StatusSendtPage } from "./_components/StatusSendtPage";
+import { StatusMottattPage } from "./_components/StatusMottattPage";
+import { StatusUnderBehandlingPage } from "./_components/StatusUnderBehandlingPage";
+import { StatusFerdigbehandletPage } from "./_components/StatusFerdigbehandletPage";
 
 export const dynamic = "force-dynamic";
