+++ src/components/soknaderList/list/soknadCard/status/MottattCard.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from "next-intl";
 import { FileTextIcon } from "@navikt/aksel-icons";
 
-import StatusCard from "../../../statusCard/StatusCard";
+import StatusCard from "@components/statusCard/StatusCard";
 
 interface Props {
     fiksDigisosId: string;
