+++ src/components/soknaderList/list/soknadCard/status/UnderBehandlingCard.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from "next-intl";
 import { PersonGavelIcon } from "@navikt/aksel-icons";
 
-import StatusCard from "../../../statusCard/StatusCard";
+import StatusCard from "@components/statusCard/StatusCard";
 
 interface Props {
     fiksDigisosId: string;
